requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-bootstrap-modal-forms==2.2.0
django-global-permissions==0.2.6
django-safedelete==1.2.2
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.4.12
install==1.3.5
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
